Добавляет валидацию заголовка и синхронизацию полей 'Тип жилья' и 'Цена за ночь'
