+++ js/render-advertisement.js
@@ -1,3 +1,8 @@
+import {wordEndings, inclineWord} from './utils.js';
+
+wordEndings['guestsEndings'];
+//wordEndings['roomsEndings'];
+
 const TYPES = {
   'palace': 'Дворец',
   'flat': 'Квартира',
@@ -55,41 +60,6 @@ const renderPhotos = (parent, cssClass, data) => {
   }
 
   parent.querySelector(cssClass).remove();
-};
-
-const wordEndings = {
-  'roomsEndings': {
-    firstState: 'комнат',
-    secondState: 'комната',
-    thirdState: 'комнаты',
-    fourthState: 'комнат',
-  },
-
-  'guestsEndings': {
-    firstState: 'гостей',
-    secondState: 'гостя',
-    thirdState: 'гостей',
-    fourthState: 'гостей',
-  },
-};
-
-const inclineWord = (num, type) => {
-  const n = num ? num.toString() : '1';
-  const last = n.slice(-1);
-  const twoLast = n.slice(-2);
-  if (twoLast === '11' || twoLast === '12' || twoLast === '13' || twoLast === '14') {
-    return `${n} ${wordEndings[type].firstState}`;
-  }
-
-  if (last === '1') {
-    return `${n} ${wordEndings[type].secondState}`;
-  }
-
-  if (last === '2' || last === '3' || last === '4') {
-    return `${n} ${wordEndings[type].thirdState}`;
-  }
-
-  return `${n} ${wordEndings[type].fourthState}`;
 };
 
 const renderCapacity = (parent, cssClass, roomsData, guestsData) => {
